Move CPM commands to command subdirectory and add new test commands
